Copy over the govuk script from govuk-guix
This script serves to tie together different scripts. It aids with
managing the $PATH, finding appropriate Ruby versions and anything
else that's useful.

It's been tweaked to be less specific to the govuk-guix repository,
whilst still keeping the current functionality.

The intention is that scripts are named govuk-*, for example
govuk-connect. You would then run them by typing:

  govuk connect