Copy over the govuk-connect script from govuk-guix
This script aims to make it easy to connect to various parts of the
GOV.UK infrastructure. It aims to provide a consistent way to do this,
even across different hosting providers.

It was originally maintained in the govuk-guix repository, as that
contained other scripts. However, govuk-guix does a lot of other
stuff, so to make this aspect less muddled, move some of the non-Guix
related scripts to a more generic repository.