Copy over the govuk-aws script from govuk-guix
Like govuk-connect, this isn't connected to Guix, but that was just a
convinient place to store the script prior to the creation of the
govuk-cli repository.

This was originally written to provide a stateless alternative to
govukcli, as well as making it easier to integrate in to scripts like
`govuk data`. Additionally, I still think it's useful to keep aroun
given it's a lot simpler compared to the similar functionality in the
govukcli or gds-cli scripts, the latter which requires installing
aws-vault as I understand it.