Stop output of aws binary location when running
`2>` redirects stderr and was causing the location of the binary to be
printed when testing for the existence of the aws binary. This can cause
issues in that the output is no longer interpreted as valid json (for
example). Instead we want to suppress the output of the location of the
binary.